fix(test): resolve nil pointer and type assertion failures in agent tests
- maintenance_test.go: initialize config and real MaintenanceTool instead of
  passing nil, which caused panic in GetTasksByFilter
- agent_financial_test.go: pass equipment_id as float64 to match handler's
  type assertion (uint was silently ignored, resulting in id=0)

+++ backend/internal/agent/service/agent_financial_test.go
@@ -87,7 +87,7 @@ func TestAgentService_HandleGetRepairCosts(t *testing.T) {
 	}
 	
 	svc := NewAgentService()
-	args := map[string]interface{}{"equipment_id": eqID}
+	args := map[string]interface{}{"equipment_id": float64(eqID)}
 	
 	result, err := svc.handleGetRepairCosts(user, args)
 	if err != nil {
